Add files via upload
Ok well, wow's api no longer links to a 3d model viewer so changed the direction into a mount display, only thing left to do is to clean up the code so that it can delete old lists for new searches and compare to a master list, plus you know CSS style the site. Enjoy the craigslist vibes atm.